requirements.txt
altgraph==0.17.4
<<<<<<< HEAD
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
idna==3.10
kiwisolver==1.4.8
Naked==0.1.32
=======
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
kiwisolver==1.4.8
matplotlib==3.10.0
>>>>>>> d33347e54785bdf917f3ee1caf18af3b22d1f604
numpy==2.2.2
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
pycryptodome==3.21.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
<<<<<<< HEAD
PyYAML==6.0.2
requests==2.32.3
shellescape==3.8.1
six==1.17.0
tkinterdnd2==0.4.2
urllib3==2.3.0
=======
six==1.17.0
>>>>>>> d33347e54785bdf917f3ee1caf18af3b22d1f604
